Move twig customization to it's own service provider
